# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Funnel AI Studio API & MCP Suite",
    "description": "API d'orchestration cognitive et suite d'outils d'élite pour concevoir, affiner la stratégie, générer des images et modifier en sandbox des tunnels de vente haute conversion clé en main avec design Luxe Studio et copywriting persuasif.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://funnelai.coolify.dallico.com/api/v1",
      "description": "Serveur Funnel AI Studio Production"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "summary": "Vérifier le statut du moteur Funnel AI Studio",
        "responses": {
          "200": {
            "description": "Statut opérationnel"
          }
        }
      }
    },
    "/templates": {
      "get": {
        "operationId": "listTemplates",
        "summary": "Lister tous les modèles de tunnels disponibles",
        "responses": {
          "200": {
            "description": "Liste des templates"
          }
        }
      }
    },
    "/funnels/analyze": {
      "post": {
        "operationId": "analyzeOffer",
        "summary": "Analyser une offre et déduire la stratégie de conversion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "request"
                ],
                "properties": {
                  "request": {
                    "type": "string",
                    "description": "Description commerciale de l'offre"
                  },
                  "templateId": {
                    "type": "string",
                    "description": "ID optionnel d'un template"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analyse business"
          }
        }
      }
    },
    "/funnels/refine-strategy": {
      "post": {
        "operationId": "refineStrategy",
        "summary": "Combler les lacunes stratégiques (tarifs, témoignages, garantie, livrables, trafic, délais)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "request"
                ],
                "properties": {
                  "request": {
                    "type": "string"
                  },
                  "identifiedGaps": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "userPreferences": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stratégie enrichie 360"
          }
        }
      }
    },
    "/funnels/generate": {
      "post": {
        "operationId": "generateCompleteFunnel",
        "summary": "Générer un tunnel de vente complet haute conversion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "request"
                ],
                "properties": {
                  "request": {
                    "type": "string"
                  },
                  "templateId": {
                    "type": "string"
                  },
                  "generateHeroImage": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tunnel complet généré"
          }
        }
      }
    },
    "/funnels/generate-image": {
      "post": {
        "operationId": "generateHeroImage",
        "summary": "Générer une image photoréaliste personnalisée pour le tunnel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string"
                  },
                  "style": {
                    "type": "string"
                  },
                  "aspectRatio": {
                    "type": "string",
                    "default": "16:9"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image générée"
          }
        }
      }
    },
    "/funnels/modify": {
      "post": {
        "operationId": "modifyFunnelSandbox",
        "summary": "Modifier le code d'un tunnel en sandbox (HTML, CSS, JS) selon une consigne",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "instruction"
                ],
                "properties": {
                  "instruction": {
                    "type": "string"
                  },
                  "indexHtml": {
                    "type": "string"
                  },
                  "stylesCss": {
                    "type": "string"
                  },
                  "scriptJs": {
                    "type": "string"
                  },
                  "standaloneHtml": {
                    "type": "string"
                  },
                  "injectedAssets": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code modifié"
          }
        }
      }
    },
    "/funnels/export": {
      "post": {
        "operationId": "exportFunnelBundle",
        "summary": "Exporter le tunnel sous forme d'archive ZIP téléchargeable ou fichier HTML",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productName": {
                    "type": "string",
                    "default": "Mon-Tunnel-Luxe"
                  },
                  "indexHtml": {
                    "type": "string"
                  },
                  "stylesCss": {
                    "type": "string"
                  },
                  "scriptJs": {
                    "type": "string"
                  },
                  "standaloneHtml": {
                    "type": "string"
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "zip",
                      "html",
                      "json"
                    ],
                    "default": "zip"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Archive ZIP ou bundle HTML généré"
          }
        }
      }
    }
  }
}